feat(cb2-4089): move spinner store components into seperate folders/files
