+++ src/app/store/spinner/actions/spinner.actions.ts
@@ -1,0 +1,3 @@
+import { createAction, props } from '@ngrx/store';
+
+export const setSpinnerState = createAction('[UI/spinner] set spinner state', props<{ showSpinner: boolean }>());
